+++ server/src-lib/Hasura/RQL/DML/Delete.hs
@@ -34,7 +34,7 @@ import           Hasura.Session
 
 validateDeleteQWith
   :: (UserInfoM m, QErrM m, TableInfoRM ('Postgres 'Vanilla) m)
-  => SessVarBldr ('Postgres 'Vanilla) m
+  => SessionVariableBuilder ('Postgres 'Vanilla) m
   -> (ColumnType ('Postgres 'Vanilla) -> Value -> m S.SQLExp)
   -> DeleteQuery
   -> m (AnnDel ('Postgres 'Vanilla))