+++ server/src-lib/Hasura/RQL/DML/Internal.hs
@@ -1,7 +1,4 @@
 module Hasura.RQL.DML.Internal where
--- ( mkAdminRolePermInfo
--- , SessVarBldr
--- ) where
 
 import           Hasura.Prelude
 
@@ -207,7 +204,11 @@ fetchRelTabInfo refTabName =
   modifyErrAndSet500 ("foreign " <> ) $
     askTabInfoSource refTabName
 
-type SessVarBldr b m = SessionVarType b -> SessionVariable -> m (SQLExpression b)
+data SessionVariableBuilder b m
+  = SessionVariableBuilder
+  { _svbCurrentSession :: !(SQLExpression b)
+  , _svbVariableParser :: !(SessionVarType b -> SessionVariable -> m (SQLExpression b))
+  }
 
 fetchRelDet
   :: (UserInfoM m, QErrM m, TableInfoRM b m, Backend b)
@@ -234,7 +235,7 @@ fetchRelDet relName refTabName = do
 checkOnColExp
   :: (UserInfoM m, QErrM m, TableInfoRM b m, Backend b)
   => SelPermInfo b
-  -> SessVarBldr b m
+  -> SessionVariableBuilder b m
   -> AnnBoolExpFldSQL b
   -> m (AnnBoolExpFldSQL b)
 checkOnColExp spi sessVarBldr annFld = case annFld of
@@ -267,7 +268,7 @@ checkOnColExp spi sessVarBldr annFld = case annFld of
 
 convAnnBoolExpPartialSQL
   :: (Applicative f, Backend backend)
-  => SessVarBldr backend f
+  => SessionVariableBuilder backend f
   -> AnnBoolExpPartialSQL backend
   -> f (AnnBoolExpSQL backend)
 convAnnBoolExpPartialSQL f =
@@ -275,7 +276,7 @@ convAnnBoolExpPartialSQL f =
 
 convAnnColumnCaseBoolExpPartialSQL
   :: (Applicative f, Backend backend)
-  => SessVarBldr backend f
+  => SessionVariableBuilder backend f
   -> AnnColumnCaseBoolExpPartialSQL backend
   -> f (AnnColumnCaseBoolExp backend (SQLExpression backend))
 convAnnColumnCaseBoolExpPartialSQL f =
@@ -283,17 +284,18 @@ convAnnColumnCaseBoolExpPartialSQL f =
 
 convPartialSQLExp
   :: (Applicative f)
-  => SessVarBldr backend f
+  => SessionVariableBuilder backend f
   -> PartialSQLExp backend
   -> f (SQLExpression backend)
-convPartialSQLExp f = \case
+convPartialSQLExp sessVarBldr = \case
   PSESQLExp sqlExp                 -> pure sqlExp
-  PSESessVar colTy sessionVariable -> f colTy sessionVariable
+  PSESession                       -> pure $ _svbCurrentSession sessVarBldr
+  PSESessVar colTy sessionVariable -> (_svbVariableParser sessVarBldr) colTy sessionVariable
 
 sessVarFromCurrentSetting
-  :: (Applicative f) => CollectableType PGScalarType -> SessionVariable -> f S.SQLExp
-sessVarFromCurrentSetting pgType sessVar =
-  pure $ sessVarFromCurrentSetting' pgType sessVar
+  :: (Applicative f) => SessionVariableBuilder ('Postgres pgKind) f
+sessVarFromCurrentSetting =
+  SessionVariableBuilder currentSession $ \ty var -> pure $ sessVarFromCurrentSetting' ty var
 
 sessVarFromCurrentSetting' :: CollectableType PGScalarType -> SessionVariable -> S.SQLExp
 sessVarFromCurrentSetting' ty sessVar =
@@ -329,7 +331,7 @@ currentSession = S.SEUnsafe "current_setting('hasura.user')::json"
 checkSelPerm
   :: (UserInfoM m, QErrM m, TableInfoRM b m, Backend b)
   => SelPermInfo b
-  -> SessVarBldr b m
+  -> SessionVariableBuilder b m
   -> AnnBoolExpSQL b
   -> m (AnnBoolExpSQL b)
 checkSelPerm spi sessVarBldr =
@@ -340,12 +342,13 @@ convBoolExp
   => FieldInfoMap (FieldInfo b)
   -> SelPermInfo b
   -> BoolExp b
-  -> SessVarBldr b m
+  -> SessionVariableBuilder b m
   -> TableName b
   -> ValueParser b m (SQLExpression b)
   -> m (AnnBoolExpSQL b)
 convBoolExp cim spi be sessVarBldr rootTable rhsParser = do
-  abe <- annBoolExp rhsParser rootTable cim $ unBoolExp be
+  let boolExpRHSParser = BoolExpRHSParser rhsParser $ _svbCurrentSession sessVarBldr
+  abe <- annBoolExp boolExpRHSParser rootTable cim $ unBoolExp be
   checkSelPerm spi sessVarBldr abe
 
 dmlTxErrorHandler :: Q.PGTxErr -> QErr